pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Bibek Timilsina <bibektimilsina000@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
seleniumbase = "^4.32.5"
uvicorn = "^0.32.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.17.0"}
sqlmodel = "^0.0.22"
pyjwt = "^2.9.0"
pydantic-settings = "^2.6.0"
pydantic = { extras = ["email"], version = "^2.9.2" }
psycopg = { extras = ["binary"], version = "^3.2.3" }
emails = "^0.6"
jinja2 = "^3.1.4"
python-multipart = "^0.0.16"
celery = "^5.4.0"
redis = "^5.2.0"
asyncpg = "^0.30.0"
alembic = "^1.13.3"
tenacity = "^9.0.0"
flower = "^2.0.1"
google-api-python-client = "^2.149.0"
oauth2client = "^4.1.3"
colorlog = "^6.9.0"
bcrypt = "4.0.1"
passlib = "^1.7.4"
fake-useragent = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
